+++ src/index.ts
@@ -3,4 +3,5 @@ export * from './api';
 export * from './converter';
 export * from './lib/decrypt';
 export * from './lib/get-url';
+export {httpAgent, httpsAgent, getBuffer, getJson, getText} from './lib/http';
 export * from './metadata-writer';